Generic code quality improvements
- Changed (almost) all C-style casts to C++-style casts
    Only exception: Task_t definitions. These will need to be addressed at a later date
	Pointless casts were removed
	IMPORTANT: there are two const_casts inside client.cpp. Const-correctness isn't respected there, FIX!
- Corrected a few entries among the C interfaces (also for const-correctness)
- Fixed indentation here and there
- Fixed float literals here and there
- Enforced usage of appropriate enums, rather than int, in a few instances
- Removed useless warning suppression from CMakeLists.txt
- Replaced some usage of float[3][4] with matrix3x4, like in the FWGS engine
- Updated GET_PRIVATE to accept a template argument, in a similar fashion to GetClassPtr
- Optimized a few sorting algorithms

+++ cl_dll/scoreboard.cpp
@@ -418,7 +418,7 @@ int CHudScoreboard::DrawPlayers( int xpos_rel, float list_slot, int nameoffset, 
 			else
 			{
 				// Highlight the killers name - overlay the background in red,  then draw the score text over it
-				FillRGBA( xpos - 5, ypos, FAR_RIGHT, ROW_GAP, 255, 0, 0, ( (float)15 * (float)( m_fLastKillTime - gHUD.m_flTime ) ) );
+				FillRGBA( xpos - 5, ypos, FAR_RIGHT, ROW_GAP, 255, 0, 0, ( 15.0f * ( float( m_fLastKillTime ) - gHUD.m_flTime ) ) );
 			}
 		}
 		else if( pl_info->thisplayer ) // if it is their name, draw it a different color
