+++ dlls/xen.cpp
@@ -95,10 +95,16 @@ void CXenPLight::Spawn( void )
 	UTIL_SetSize( pev, Vector( -80, -80, 0 ), Vector( 80, 80, 32 ) );
 	SetActivity( ACT_IDLE );
 	pev->nextthink = gpGlobals->time + 0.1;
-	pev->frame = RANDOM_FLOAT( 0, 255 );
+	pev->frame = RANDOM_FLOAT( 0.0f, 255.0f );
 
 	m_pGlow = CSprite::SpriteCreate( XEN_PLANT_GLOW_SPRITE, pev->origin + Vector(0,0,(pev->mins.z+pev->maxs.z)*0.5), FALSE );
-	m_pGlow->SetTransparency( kRenderGlow, (int)pev->rendercolor.x, (int)pev->rendercolor.y, (int)pev->rendercolor.z, (int)pev->renderamt, (int)pev->renderfx );
+	m_pGlow->SetTransparency( kRenderGlow,
+		static_cast<int>( pev->rendercolor.x ),
+		static_cast<int>( pev->rendercolor.y ),
+		static_cast<int>( pev->rendercolor.z ),
+		static_cast<int>( pev->renderamt ),
+		static_cast<int>( pev->renderfx )
+	);
 	m_pGlow->SetAttachment( edict(), 1 );
 }
 
@@ -186,14 +192,14 @@ void CXenHair::Spawn( void )
 
 	if( !( pev->spawnflags & SF_HAIR_SYNC ) )
 	{
-		pev->frame = RANDOM_FLOAT( 0, 255 );
-		pev->framerate = RANDOM_FLOAT( 0.7, 1.4 );
+		pev->frame = RANDOM_FLOAT( 0.0f, 255.0f );
+		pev->framerate = RANDOM_FLOAT( 0.7f, 1.4f );
 	}
 	ResetSequenceInfo();
 
 	pev->solid = SOLID_NOT;
 	pev->movetype = MOVETYPE_NONE;
-	pev->nextthink = gpGlobals->time + RANDOM_FLOAT( 0.1, 0.4 );	// Load balance these a bit
+	pev->nextthink = gpGlobals->time + RANDOM_FLOAT( 0.1f, 0.4f );	// Load balance these a bit
 }
 
 void CXenHair::Think( void )
@@ -284,8 +290,8 @@ void CXenTree::Spawn( void )
 	UTIL_SetSize( pev, Vector( -30, -30, 0 ), Vector( 30, 30, 188 ) );
 	SetActivity( ACT_IDLE );
 	pev->nextthink = gpGlobals->time + 0.1;
-	pev->frame = RANDOM_FLOAT( 0, 255 );
-	pev->framerate = RANDOM_FLOAT( 0.7, 1.4 );
+	pev->frame = RANDOM_FLOAT( 0.0f, 255.0f );
+	pev->framerate = RANDOM_FLOAT( 0.7f, 1.4f );
 
 	Vector triggerPosition;
 	UTIL_MakeVectorsPrivate( pev->angles, triggerPosition, NULL, NULL );
@@ -329,7 +335,7 @@ void CXenTree::Attack( void )
 	if( GetActivity() == ACT_IDLE )
 	{
 		SetActivity( ACT_MELEE_ATTACK1 );
-		pev->framerate = RANDOM_FLOAT( 1.0, 1.4 );
+		pev->framerate = RANDOM_FLOAT( 1.0f, 1.4f );
 		EMIT_SOUND_ARRAY_DYN( CHAN_WEAPON, pAttackMissSounds );
 	}
 }
@@ -384,7 +390,7 @@ void CXenTree::Think( void )
 		if( m_fSequenceFinished )
 		{
 			SetActivity( ACT_IDLE );
-			pev->framerate = RANDOM_FLOAT( 0.6, 1.4 );
+			pev->framerate = RANDOM_FLOAT( 0.6f, 1.4f );
 		}
 		break;
 	default:
@@ -497,7 +503,7 @@ void CXenSporeLarge::Spawn( void )
 	UTIL_MakeVectorsPrivate( pev->angles, forward, right, NULL );
 
 	// Rotate the leg hulls into position
-	for( int i = 0; i < (int)ARRAYSIZE( m_hullSizes ); i++ )
+	for( int i = 0; i < ARRAYSIZE( m_hullSizes ); i++ )
 		CXenHull::CreateHull( this, Vector( -12, -12, 0 ), Vector( 12, 12, 120 ), ( m_hullSizes[i].x * forward ) + ( m_hullSizes[i].y * right ) );
 }
 
@@ -512,10 +518,10 @@ void CXenSpore :: Spawn( void )
 
 	//SetActivity( ACT_IDLE );
 	pev->sequence = 0;
-	pev->frame = RANDOM_FLOAT( 0, 255 );
-	pev->framerate = RANDOM_FLOAT( 0.7, 1.4 );
+	pev->frame = RANDOM_FLOAT( 0.0f, 255.0f );
+	pev->framerate = RANDOM_FLOAT( 0.7f, 1.4f );
 	ResetSequenceInfo();
-	pev->nextthink = gpGlobals->time + RANDOM_FLOAT( 0.1, 0.4 );	// Load balance these a bit
+	pev->nextthink = gpGlobals->time + RANDOM_FLOAT( 0.1f, 0.4f );	// Load balance these a bit
 }
 
 const char *CXenSpore::pModelNames[] =
